requirements.txt
fastapi
uvicorn
SQLAlchemy
jinja2
python-multipart
python-slugify
feedparser
ollama
requests
beautifulsoup4
lxml
psycopg2-binary
apscheduler
chromadb
sentence-transformers
pydantic[email]
bcrypt
python-jose[cryptography]
markdown
bleach
